+++ 2DChart/Domain/Function/FunctionDto.cs
@@ -9,7 +9,7 @@ namespace _2DChart.Domain.Function
     {
         public double Min { get; set; }
         public double Max { get; set; }
-        public double Approximation { get; set; }
+        public int Approximation { get; set; }
         public string Name { get; set; }
         public string FunctionString { get; set; }
         public DateTime CreationDate { get; set; }
